requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.0.9
narwhals==1.38.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
protobuf==5.29.4
psutil==6.1.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
soupsieve==2.7
supabase==2.3.0
supabase-py==0.0.2
postgrest==0.13.0
gotrue==2.1.0
python-dotenv==1.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
yfinance==0.2.58
wrapt==1.17.2
beautifulsoup4==4.13.4
blinker==1.9.0
gradio>=3.0